// ui/src/generated/types/alertMessage.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Mercury
 * Control and manages real Smartphone devices from browser and restful apis
 * OpenAPI spec version: 1.4.7
 */
import type { AlertMessageLevel } from './alertMessageLevel'

export interface AlertMessage {
  /** Enable or disablee the alert message */
  activation?: string
  /** Alert message text to display */
  data?: string
  /** Alert message level */
  level?: AlertMessageLevel
}